[REF] l10n_be_vat_reports: split vat listing model files
