+++ l10n_be_vat_reports/wizard/__init__.py
@@ -1,2 +1,4 @@
-from . import l10n_be_partner_vat_listing
 from . import l10n_be_vat_intra
+from . import partner_vat
+from . import partner_vat_list
+from . import vat_listing_clients
